requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
iniconfig==2.0.0
install==1.3.5
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.8
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
